+++ src/idlib/Str.cpp
@@ -182,7 +182,7 @@ void idStr::operator=(const char *text)
 		return;
 	}
 
-	l = strlen(text);
+	l = (int)strlen(text);
 	EnsureAlloced(l + 1, false);
 	strcpy(data, text);
 	len = l;
@@ -201,7 +201,7 @@ int idStr::FindChar(const char *str, const char c, int start, int end)
 
 	if (end == -1)
 	{
-		end = strlen(str) - 1;
+		end = (int)strlen(str) - 1;
 	}
 	for (i = start; i <= end; i++)
 	{
@@ -226,9 +226,9 @@ int idStr::FindText(const char *str, const char *text, bool casesensitive, int s
 
 	if (end == -1)
 	{
-		end = strlen(str);
+		end = (int)strlen(str);
 	}
-	l = end - strlen(text);
+	l = end - (int)strlen(text);
 	for (i = start; i <= l; i++)
 	{
 		if (casesensitive)
@@ -568,7 +568,7 @@ void idStr::StripLeading(const char *string)
 {
 	int l;
 
-	l = strlen(string);
+	l = (int)strlen(string);
 	if (l > 0)
 	{
 		while (!Cmpn(string, l))
@@ -588,7 +588,7 @@ bool idStr::StripLeadingOnce(const char *string)
 {
 	int l;
 
-	l = strlen(string);
+	l = (int)strlen(string);
 	if ((l > 0) && !Cmpn(string, l))
 	{
 		memmove(data, data + l, len - l + 1);
@@ -623,7 +623,7 @@ void idStr::StripTrailing(const char *string)
 {
 	int l;
 
-	l = strlen(string);
+	l = (int)strlen(string);
 	if (l > 0)
 	{
 		while ((len >= l) && !Cmpn(string, data + len - l, l))
@@ -643,7 +643,7 @@ bool idStr::StripTrailingOnce(const char *string)
 {
 	int l;
 
-	l = strlen(string);
+	l = (int)strlen(string);
 	if ((l > 0) && (len >= l) && !Cmpn(string, data + len - l, l))
 	{
 		len -= l;
@@ -664,8 +664,8 @@ int idStr::Replace(const char *old, const char *nw)
 	int oldLen, newLen, i, j, count;
 	idStr oldString(data);
 
-	oldLen = strlen(old);
-	newLen = strlen(nw);
+	oldLen = (int)strlen(old);
+	newLen = (int)strlen(nw);
 
 	// Work out how big the new string will be
 	count = 0;
@@ -699,7 +699,7 @@ int idStr::Replace(const char *old, const char *nw)
 			}
 		}
 		data[j] = 0;
-		len = strlen(data);
+		len = (int)strlen(data);
 	}
 	return iReplaced;
 }
@@ -973,7 +973,7 @@ void idStr::AppendPath(const char *text)
 	if (text && text[i])
 	{
 		pos = len;
-		EnsureAlloced(len + strlen(text) + 2);
+		EnsureAlloced(len + (int)strlen(text) + 2);
 
 		if (pos)
 		{
@@ -1645,7 +1645,7 @@ void idStr::Append(char *dest, int size, const char *src)
 {
 	int l1;
 
-	l1 = strlen(dest);
+	l1 = (int)strlen(dest);
 	if (l1 >= size)
 	{
 		idLib::common->Error("idStr::Append: already overflowed");
